エラーユーティリティ用ファイルを作成
