+++ src/api/index.ts
@@ -1,7 +1,7 @@
 import { Hono } from "hono";
 import { logger } from "hono/logger";
 import routes from './resources'
-import { unknownError } from "./resources/posts/features/errors";
+import { unknownError } from "./lib/errors";
 
 const app = new Hono({ 
   strict: false // 最後の/を区別しない @see https://hono.dev/docs/api/hono#strict-mode
@@ -14,7 +14,7 @@ const app = new Hono({
   })
   .onError((error, c) => {
     console.log('Unhandled Error:', error)
-    return c.json(unknownError) // 便利だし使っちゃおう
+    return c.json(unknownError)
   })
 
 export type AppType = typeof routes
