+++ src/api/resources/posts/features/errors.ts
@@ -1,16 +1,4 @@
-import { StatusCode } from "hono/utils/http-status"
-
-export type CustomError = Readonly<{
-  code: string,
-  status: StatusCode,
-  message: string
-}>
-
-type CustomErrorInit = Readonly<{
-  names: string[]
-  message: string,
-  status?: StatusCode
-}>
+import { CustomErrorInit, createGetError } from "@/api/lib/errors"
 
 const postErrors: readonly CustomErrorInit[] = [
   {
@@ -29,19 +17,4 @@ const postErrors: readonly CustomErrorInit[] = [
   }
 ]
 
-export const unknownError = {
-  code: 'UnknownError',
-  status: 500,
-  message: '予期せぬエラーが発生しました'
-} as const satisfies CustomError
-
-export function getPostError(code: string) {
-  const error = postErrors.find(e => e.names.includes(code))
-  if (!error) return null
-
-  return {
-    code,
-    status: error.status ?? 400,
-    message: error.message
-  } satisfies  CustomError
-}
+export const getPostError = createGetError(postErrors)
